Moved helpers to collections.py
